👌 IMPROVE: Extract utils & validation from src of /api
