+++ packages/api/src/routers/plan.ts
@@ -5,9 +5,9 @@ import { z } from "zod";
 import type { Plan } from "@plan-prise/db-prisma";
 import PP_Error from "@plan-prise/errors";
 
+import { planSettingsSchema } from "../../validation/plan";
 import { MUTATION_SUCCESS, PLAN_NEW } from "../constants";
 import { authProcedure, createTRPCRouter } from "../trpc";
-import { planSettingsSchema } from "../validation/plan";
 
 const addMedic = (medicId: string, medicsOrder: Plan["medicsOrder"]) => {
   if (isCuid(medicId)) {
